+++ tests/unit/specs/src/containers/ConfirmationContainer/components/ConfirmSignModal/ConfirmSignModal.spec.js
@@ -2,7 +2,7 @@ import VueX from 'vuex';
 import { shallowMount } from '@vue/test-utils';
 import ConfirmSignModal from '@/containers/ConfirmationContainer/components/ConfirmSignModal/ConfirmSignModal.vue';
 import VueQrcode from '@xkeshi/vue-qrcode';
-
+import { state, getters } from '@@/helpers/mockStore';
 import { Tooling } from '@@/helpers';
 const confirmSignMessage = jest.fn();
 
@@ -31,20 +31,9 @@ describe('ConfirmSignModal.vue', () => {
     i18n = baseSetup.i18n;
     store = baseSetup.store;
 
-    const wallet = {
-      getChecksumAddressString: function() {
-        return '0xDECAF9CD2367cdbb726E904cD6397eDFcAe6068D';
-      }
-    };
-
-    const getters = {
-      wallet: () => {
-        return wallet;
-      }
-    };
-
     store = new VueX.Store({
-      getters
+      getters,
+      state
     });
   });