+++ tests/unit/specs/src/dapps/RegisterDomainContainer/containers/EnsBidContainer/EnsBidContainer.spec.js
@@ -3,11 +3,9 @@ import Vuex from 'vuex';
 import { shallowMount } from '@vue/test-utils';
 import EnsBidContainer from '@/dapps/RegisterDomain/containers/EnsBidContainer/EnsBidContainer.vue';
 import JsonStringModal from '@/dapps/RegisterDomain/components/JsonStringModal/JsonStringModal.vue';
-import nodeList from '@/networks';
-import url from 'url';
-import Web3 from 'web3';
 import sinon from 'sinon';
 import { Misc } from '@/helpers';
+import { state, getters } from '@@/helpers/mockStore';
 
 import { Tooling } from '@@/helpers';
 
@@ -48,27 +46,9 @@ describe('EnsBidContainer.vue', () => {
     i18n = baseSetup.i18n;
     store = baseSetup.store;
 
-    const network = nodeList['ETH'][3];
-    const hostUrl = url.parse(network.url);
-
-    const newWeb3 = new Web3(
-      `${hostUrl.protocol}//${hostUrl.hostname}:${network.port}${
-        hostUrl.pathname
-      }`
-    );
-
-    const getters = {
-      web3: () => {
-        return newWeb3;
-      }
-    };
-
     store = new Vuex.Store({
-      getters
-    });
-
-    store.replaceState({
-      web3: newWeb3
+      getters,
+      state
     });
 
     Vue.config.errorHandler = () => {};