+++ tests/unit/specs/src/layouts/InterfaceLayout/components/CurrencyPicker/CurrencyPicker.spec.js
@@ -1,10 +1,8 @@
 import Vuex from 'vuex';
 import { shallowMount } from '@vue/test-utils';
 import CurrencyPicker from '@/layouts/InterfaceLayout/components/CurrencyPicker/CurrencyPicker.vue';
-import nodeList from '@/networks';
-import url from 'url';
-import Web3 from 'web3';
 import { Tooling } from '@@/helpers';
+import { state, getters } from '@@/helpers/mockStore';
 
 const currency = [
   { symbol: 'BTC', name: 'Bitcoin' },
@@ -20,26 +18,9 @@ describe('CurrencyPicker.vue', () => {
     i18n = baseSetup.i18n;
     store = baseSetup.store;
 
-    const network = nodeList['ETH'][3];
-    const hostUrl = url.parse(network.url);
-
-    const newWeb3 = new Web3(
-      `${hostUrl.protocol}//${hostUrl.hostname}:${network.port}${
-        hostUrl.pathname
-      }`
-    );
-
-    const getters = {
-      web3: () => {
-        return newWeb3;
-      },
-      network: () => {
-        return network;
-      }
-    };
-
     store = new Vuex.Store({
-      getters
+      getters,
+      state
     });
   });