+++ CHANGELOG.md
@@ -15,6 +15,10 @@
 
 - Fix deploy contract params and validation issue [#502](https://github.com/MyEtherWallet/MyEtherWallet/pull/502)
 
+### Feature
+
+- Componentize wallet options, change all wallet address instance to coinbase [](<>)
+
 ### Release v5.0.0-beta.2
 
 ### Hotfix